refactor: modularize dashboard and league services, move types to shared directory, and implement server-side data prefetching for improved performance
